Shifted to Pavel's Debian build system
I have changed the current Debian build system to the one made by
Pavel Shramov. It is a lot simpler than the current one. Also it is
similar to what can normally be found in the Debian archieve, and I
therefore assume that it will be easier to some day get an offecial
WxHaskell Debian package.

All of the files in the debian directory is the work of Pavel.

I changed only one thing, as compared to Pavel's version. Namely that
I removed the dependcy on libghc6-time-dev. I did it, as this
dependency makes the package un-buildable on Debian unstable (GHC
6.8). At release 0.10.3, some users requested a GHC 6.8 version, and
it is therefore important that we can provide one.

darcs-hash:20080413131752-73052-fcd3cccb4765f71c7a564f0e96ff7dcdbd9ecfd8.gz